requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
dvc==3.22.0
xgboost==1.7.6
gunicorn==21.2.0
Flask==2.3.3
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
numpy==1.24.3
ucimlrepo==0.0.7